♻️[Refactor]#98 - 리팩에 따른 코드 수정
