+++ DanceMachine/DanceMachine/Common/Components/Feedback/FeedbackCard.swift
@@ -149,7 +149,6 @@ struct FeedbackCard: View {
           .truncationMode(.tail)
         }
       }
-      //      Spacer()
     }
   }
